+++ apps/hash-frontend/src/pages/shared/entities-visualizer/entities-table-data.ts
