+++ apps/hash-frontend/src/pages/shared/entities-visualizer/grid-view.tsx
