+++ apps/hash-frontend/src/pages/shared/entities-visualizer/grid-view/grid-view-item-skeleton.tsx
