+++ apps/hash-frontend/src/pages/shared/entities-visualizer/header.tsx
@@ -1,0 +1,6 @@
+export { FilterRibbon } from "./header/filter-ribbon";
+export { QueryCount } from "./header/query-count";
+export {
+  VisualizerHeader,
+  visualizerHeaderHeight,
+} from "./header/visualizer-header";
