+++ apps/hash-frontend/src/pages/shared/entities-visualizer/shared/property-filters/get-operators-for-kind.ts
@@ -1,4 +1,7 @@
-import type { FilterValueKind, PropertyFilterOperator } from "./types";
+import type {
+  FilterValueKind,
+  PropertyFilterOperator,
+} from "./property-filter";
 
 export type OperatorDescriptor = {
   operator: PropertyFilterOperator;
