+++ apps/hash-frontend/src/pages/shared/entities-visualizer/shared/property-filters/property-filter.ts
