+++ apps/hash-frontend/src/pages/shared/entities-visualizer/shared/traversal-paths.ts
