+++ apps/hash-frontend/src/pages/shared/text-icon-cell.ts
@@ -5,10 +5,10 @@ import { customColors } from "@hashintel/design-system/theme";
 import {
   getCellHorizontalPadding,
   getYCenter,
-} from "../../../../components/grid/utils";
-import { drawTextWithIcon } from "../../../../components/grid/utils/draw-text-with-icon";
+} from "../../components/grid/utils";
+import { drawTextWithIcon } from "../../components/grid/utils/draw-text-with-icon";
 
-import type { CustomIcon } from "../../../../components/grid/utils/custom-grid-icons";
+import type { CustomIcon } from "../../components/grid/utils/custom-grid-icons";
 import type { CustomCell, CustomRenderer } from "@glideapps/glide-data-grid";
 
 export interface TextIconCellProps {
