+++ apps/hash-frontend/src/pages/shared/types-table.tsx
@@ -34,9 +34,9 @@ import {
 } from "../../shared/use-actors";
 import { useAuthenticatedUser } from "./auth-info-context";
 import { createRenderChipCell } from "./chip-cell";
-import { createRenderTextIconCell } from "./entities-visualizer/entities-table/text-icon-cell";
 import { useSlideStack } from "./slide-stack";
 import { TableHeaderToggle } from "./table-header-toggle";
+import { createRenderTextIconCell } from "./text-icon-cell";
 import { TOP_CONTEXT_BAR_HEIGHT } from "./top-context-bar";
 import { TypeGraphVisualizer } from "./type-graph-visualizer";
 import { visualizerViewIcons } from "./visualizer-views";
@@ -44,7 +44,7 @@ import { visualizerViewIcons } from "./visualizer-views";
 import type { CustomIcon } from "../../components/grid/utils/custom-grid-icons";
 import type { FilterState } from "../../shared/table-header";
 import type { ChipCell } from "./chip-cell";
-import type { TextIconCell } from "./entities-visualizer/entities-table/text-icon-cell";
+import type { TextIconCell } from "./text-icon-cell";
 import type { VisualizerView } from "./visualizer-views";
 import type {
   Item,
